moved NewValidation Resource to Cluster Validation resource, updated resource paths in wsapp.py, pulled out some mess into utity class
