+++ app/ws/misc_utilities/request_parsers.py
@@ -20,3 +20,11 @@ class RequestParsers:
         parser.add_argument('level', help="Validation message levels", location="args")
         parser.add_argument('static_validation_file', help="Use pre-generated validations", location="args")
         return parser
+
+    @staticmethod
+    def cluster_validation_parser():
+        parser = reqparse.RequestParser()
+        parser.add_argument('section', help="Validation section", location="args")
+        parser.add_argument('level', help="Validation message levels", location="args")
+        parser.add_argument('force_run', help="Force the validation to run", location="args")
+        return parser
