+++ app/ws/services/validation_service.py
@@ -1,10 +1,13 @@
 import json
 import os
 import logging
+import re
 
 from ws.db_connection import override_validations
-from ws.validation_dir.validations_utils import PermissionsObj, ValidationParams
-from ws.validation import is_newer_files, update_val_schema_files, validate_study
+from ws.validation_dir.validations_utils import PermissionsObj, ValidationParams, ClusterValidationParams, \
+    ValidationUtils
+from ws.validation import is_newer_files, update_val_schema_files, validate_study, submitJobToCluster, job_status, \
+    is_newer_timestamp
 
 logger = logging.getLogger('wslog')
 
@@ -118,3 +121,72 @@ class ValidationService:
             return {"error": msg}
 
         return {"success": val_feedback}
+
+    def cluster(self, val_params: ClusterValidationParams, perms: PermissionsObj, script):
+        """
+        Either submits a job to the cluster, or returns the existing validation file. If a validation job is already
+        running for the study, it will break and return a message to the user informing them that it is still running.
+        We also check if there have been any updates to the study directory before submitting a new cluster job.
+
+        :param val_params: ClusterValidationParams object containing the validation params such as section.
+        :param perms: PermissionsObj
+        """
+
+        logger.info("Validation params are - " + str(val_params.log_category) + " " + str(val_params.section))
+        para = ' -l {level} -i {study_id} -u {token} -s {section}'.format(
+                                                                    level=val_params.log_category,
+                                                                    study_id=perms.study_id,
+                                                                    token=perms.user_token,
+                                                                    section=val_params.section)
+
+        file_name = ValidationUtils.find_validation_report_filename(val_params, perms)
+
+        if file_name:
+            # If we have a file_name, this means a job may already be in progress. We want to check the status of that
+            # job before doing anything else.
+            result = file_name[:-5].split('_')
+            sub_job_id = result[2]
+            # bacct -l 3861194
+            status = job_status(sub_job_id)
+            logger.info("job status " + sub_job_id + " " + status)
+
+            if status == "PEND" or status == "RUN":
+                return {
+                    "message": "Validation is already in progress. Job " + sub_job_id + " is in running or pending state"}
+
+            file_name = perms.study_location + "/" + file_name
+            if os.path.isfile(file_name) and status == "DONE":
+                if not val_params.force_run:
+                    # if we don't want to force the validation to run for whatever reason, try and retrieve the existing
+                    # validation schema.
+                    return ValidationUtils.load_validation_file(file_name)
+                else:
+                    if is_newer_timestamp(perms.study_location, file_name):
+                        # if the study folder has been updated, we want to delete the validation file and run a new job.
+                        os.remove(file_name)
+                        command = script + ' ' + para
+                        return submitJobToCluster(command, val_params.section, perms.study_location)
+                    else:
+                        return ValidationUtils.load_validation_file(file_name)
+
+            elif os.path.isfile(file_name) and os.path.getsize(file_name) > 0:
+                if is_newer_timestamp(perms.study_location, file_name):
+                    logger.info("Job status is not present but update in study directory detected, creating new job")
+                    os.remove(file_name)
+                    command = script + ' ' + para
+                    return submitJobToCluster(command, val_params.section, perms.study_location)
+                else:
+                    logger.info(
+                        "Job status is not present and no update detected in the study directory, returning"
+                        " validation file."
+                    )
+                    return ValidationUtils.load_validation_file(file_name)
+        else:
+            try:
+                os.remove(file_name)
+            except Exception as e:
+                pass
+                # submit a new job return job id
+                logger.info("No validation file present , creating new job")
+                command = script + ' ' + para
+                return submitJobToCluster(command, val_params.section, perms.study_location)
